backend/requirements.txt
fastapi
uvicorn
paho-mqtt
influxdb-client
pydantic
pytest
httpx
python-dotenv
